# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ROBOTFORGE API
  description: Robot Data Infrastructure for Physical AI
  version: 0.1.0
  contact:
    name: Mohammed Maheer
    email: maheer@accelerationrobotics.com
servers:
  - url: https://api.robotforge.io/v1
    description: Production
  - url: http://localhost:3000/api
    description: Local development
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, name]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                name: { type: string }
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /auth/login:
    post:
      tags: [Auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
...
  /auth/logout:
    delete:
      tags: [Auth]
      summary: Logout and invalidate refresh token
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logged out

  # ── Episodes ───────────────────────────────────────────
...
  /episodes:
    get:
      tags: [Episodes]
      summary: List episodes (paginated)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20, maximum: 100 }
        - name: task
          in: query
          schema: { type: string }
        - name: embodiment
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Paginated episode list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Episode'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
  /episodes/{id}:
    get:
      tags: [Episodes]
      summary: Get episode detail with download URL
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Episode details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Episode'

  # ── Robots (ROS 2 bridge) ─────────────────────────────
  /robots:
    get:
      tags: [Robots]
      summary: List connected robots
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Connected robots
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ConnectedRobot'
...
  /fleet/status:
    get:
      tags: [Fleet]
      summary: Fleet-wide status (all robots, namespaces, DDS graph)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Fleet status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FleetStatus'
  /fleet/robots:
    get:
      tags: [Fleet]
      summary: List all fleet robots with ROS 2 node health
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Fleet robots
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/FleetRobot'
  /fleet/robots/{id}/namespace:
    post:
      tags: [Fleet]
      summary: Assign robot to ROS 2 namespace
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [namespace]
              properties:
                namespace: { type: string, description: "ROS 2 namespace e.g. /robot/arm1" }
      responses:
        '200':
          description: Namespace assigned

  # ── Recording (rosbag2/MCAP) ───────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AuthResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            accessToken: { type: string }
            refreshToken: { type: string }
            expiresIn: { type: integer }
            user:
              $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id: { type: string }
        email: { type: string, format: email }
        name: { type: string }
        role: { type: string, enum: [operator, developer, enterprise, admin] }
        tier: { type: string, enum: [free, starter, professional, enterprise] }
        createdAt: { type: string, format: date-time }
...
    Episode:
      type: object
      properties:
        id: { type: string }
        sessionId: { type: string }
        robotId: { type: string }
        embodiment: { type: string }
        task: { type: string }
        durationMs: { type: integer }
        frameCount: { type: integer }
        qualityScore: { type: number }
        status: { type: string, enum: [recording, processing, packaged, listed, failed] }
        sensorModalities: { type: array, items: { type: string } }
        createdAt: { type: string, format: date-time }
...
    ConnectedRobot:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        embodiment: { type: string }
        connectionType: { type: string }
        ipAddress: { type: string }
        status: { type: string, enum: [connected, disconnected, error, recording] }
        batteryLevel: { type: integer }
        cameras:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              name: { type: string }
              fps: { type: integer }
              resolution:
                type: object
                properties:
                  width: { type: integer }
                  height: { type: integer }
...
    Pagination:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        totalPages: { type: integer }

    # ── Fleet Schemas ──────────────────────────────────────
    FleetStatus:
      type: object
      properties:
        totalRobots: { type: integer }
        activeRobots: { type: integer }
        namespaces:
          type: array
          items: { type: string }
        ddsGraphHealthy: { type: boolean }
        robots:
          type: array
          items:
            $ref: '#/components/schemas/FleetRobot'
    FleetRobot:
      type: object
      properties:
        robotId: { type: string }
        name: { type: string }
        namespace: { type: string, description: "ROS 2 namespace e.g. /robot/arm1" }
        embodiment: { type: string }
        connectionType: { type: string, enum: [ros2, grpc, websocket, usb] }
        status: { type: string, enum: [connected, disconnected, error, recording] }
        ros2Status:
          type: object
          properties:
            nodeActive: { type: boolean }
            controllerState: { type: string, enum: [inactive, active, unconfigured] }
            ddsConnected: { type: boolean }
            topicHz:
              type: object
              additionalProperties: { type: number }
            lastHeartbeat: { type: string, format: date-time }

    # ── Recording Schemas ──────────────────────────────────
...